Inline decision view query hooks
